pyproject.toml
[tool.poetry]
name = "qualitative-coding"
version = "1.4.0"
description = "Qualitative coding tools for computer scientists"
authors = ["Chris Proctor <chris@chrisproctor.net>"]
documentation = "https://qualitative-coding.readthedocs.io"
license = "AGPL3"
readme = "README.md"
packages = [{include = "qualitative_coding"}]
include = ["refi_qda/schema.xml"]

[tool.poetry.urls]
repository = "https://github.com/cproctor/qualitative-coding"

[tool.poetry.dependencies]
python = ">=3.9"
tqdm = "^4.66.3"
tabulate = "^0.9.0"
numpy = "^1.26.0"
pyyaml = "^6.0.1"
click = "^8.1.7"
sqlalchemy = "^2.0.21"
semver = "^3.0.2"
xmlschema = "^3.0.2"
structlog = "^24.4.0"
rich = "^13.7.1"
more-itertools = "^10.4.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.3.7"
sphinx-rtd-theme = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
qc = "qualitative_coding.cli:cli"

